+++ parse-gwt/src/test/java/org/parseplatform/client/SimpleBean.java
@@ -1,0 +1,44 @@
+package org.parseplatform.client;
+
+import org.parseplatform.JSONMapper;
+
+@JSONMapper
+public class SimpleBean {
+
+    private int id;
+    private String name;
+    private double balance;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public double getBalance() {
+        return balance;
+    }
+
+    public void setBalance(double balance) {
+        this.balance = balance;
+    }
+
+    @Override
+    public String toString() {
+        return "SimpleBean{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", balance=" + balance +
+                '}';
+    }
+}