+++ processor/src/main/java/org/parseplatform/JSONMapper.java
@@ -1,0 +1,4 @@
+package org.parseplatform;
+
+public @interface JSONMapper {
+}